完成infer，对于infer后得到的outputs进行整理成一个容器，其中是每个anchor对应的box坐标，最高score和最高score对应的类别index。

+++ yolov8_tensorrt/parameters.h
@@ -17,5 +17,5 @@ typedef struct Parameters {
 struct alignas(float) Detection {
 	float box[4];
 	float conf;
-	float class_id;
+	int class_id;
 };
